+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Changed
+
+- Simplified exception hierarchy and standardized interfaces for all `qcop` exceptions.
+- Wrapped all external program errors, like `qcparse` errors, in `ExternalProgramError` so that we can maintain a clean interface for upstream callers, like `ChemCloud`, to manage exceptions.
+
 ## [0.9.8] - 2025-02-28
 
 ### Added